added mjolnir - jormungandr load tool
