+++ testing/mjolnir/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "mjolnir"
+version = "0.1.0"
+authors = ["dkijania <dariusz.kijania@iohk.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+jormungandr-lib = { path = "../../jormungandr-lib" }
+jortestkit = { git = "https://github.com/input-output-hk/jortestkit.git", branch = "master" }
+jormungandr-testing-utils = { path = "../jormungandr-testing-utils" }
+jormungandr-integration-tests = { path = "../jormungandr-integration-tests" }
+thiserror = "1.0"
+structopt = "^0.3"
+assert_fs = "1.0"
+indicatif = "0.15"
+chain-impl-mockchain = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = [ "property-test-api" ] }
+chain-addr      = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = [ "property-test-api" ] }
+chain-core      = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
+chain-crypto    = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = [ "property-test-api" ] }
+chain-time      = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
+chain-storage   = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = ["with-bench"] }
+chain-vote      = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
+tokio = { version = "0.2", features = ["macros"] }
+
+[build-dependencies]
+versionisator = "1.0.2"
