+++ testing/mjolnir/src/main.rs
@@ -1,0 +1,20 @@
+mod mjolnir_app;
+
+use std::error::Error;
+use structopt::StructOpt;
+
+fn main() {
+    mjolnir_app::Mjolnir::from_args()
+        .exec()
+        .unwrap_or_else(report_error)
+}
+
+fn report_error(error: Box<dyn Error>) {
+    eprintln!("{}", error);
+    let mut source = error.source();
+    while let Some(sub_error) = source {
+        eprintln!("  |-> {}", sub_error);
+        source = sub_error.source();
+    }
+    std::process::exit(1)
+}
