+++ testing/mjolnir/src/mjolnir_app/bootstrap/config.rs
@@ -1,32 +1,11 @@
-mod scenario;
-
-use crate::common::jormungandr::{JormungandrError, StartupError};
+use crate::mjolnir_app::{
+    bootstrap::scenario::{DurationBasedClientLoad, IterationBasedClientLoad},
+    MjolnirError,
+};
 use chain_impl_mockchain::key::Hash;
 use jormungandr_lib::interfaces::TrustedPeer;
-use jormungandr_testing_utils::testing::node::RestError;
-pub use scenario::{DurationBasedClientLoad, IterationBasedClientLoad};
 use std::path::PathBuf;
 use std::result::Result;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum ClientLoadError {
-    #[error("cannot spawn not with version '{0}', looks like it's incorrect one")]
-    VersionNotFound(String),
-    #[error("cannot find node with alias(0). Please run 'describe' command ")]
-    NodeAliasNotFound(String),
-    #[error("cannot query rest")]
-    RestError(#[from] RestError),
-    #[error("cannot bootstrap node")]
-    StartupError(#[from] StartupError),
-    #[error("jormungandr error")]
-    JormungandrError(#[from] JormungandrError),
-    #[error("node client error")]
-    InternalClientError,
-    #[error("pace is too low ({0})")]
-    PaceTooLow(u64),
-}
-
 pub struct PassiveBootstrapLoad {
     config: ClientLoadConfig,
 }
@@ -36,9 +15,9 @@ impl PassiveBootstrapLoad {
         Self { config }
     }
 
-    pub fn exec(self, scenario: ScenarioType) -> Result<(), ClientLoadError> {
+    pub fn exec(self, scenario: ScenarioType) -> Result<(), MjolnirError> {
         if self.config.pace() < 2 {
-            return Err(ClientLoadError::PaceTooLow(2));
+            return Err(MjolnirError::PaceTooLow(2));
         }
 
         match scenario {
@@ -100,5 +79,21 @@ impl ClientLoadConfig {
 
     pub fn pace(&self) -> u64 {
         self.pace
+    }
+
+    pub fn block0_hash(&self) -> &Hash {
+        &self.block0_hash
+    }
+
+    pub fn measure(&self) -> bool {
+        self.measure
+    }
+
+    pub fn count(&self) -> u32 {
+        self.count
+    }
+
+    pub fn initial_storage(&self) -> &Option<PathBuf> {
+        &self.initial_storage
     }
 }
