+++ testing/mjolnir/src/mjolnir_app/error.rs
@@ -1,0 +1,17 @@
+use jormungandr_integration_tests::common::jormungandr::{JormungandrError, StartupError};
+use jormungandr_testing_utils::testing::node::RestError;
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum MjolnirError {
+    #[error("cannot query rest")]
+    RestError(#[from] RestError),
+    #[error("cannot bootstrap node")]
+    StartupError(#[from] StartupError),
+    #[error("jormungandr error")]
+    JormungandrError(#[from] JormungandrError),
+    #[error("node client error")]
+    InternalClientError,
+    #[error("pace is too low ({0})")]
+    PaceTooLow(u64),
+}
