+++ testing/mjolnir/src/mjolnir_app/explorer/mod.rs
@@ -1,32 +1,26 @@
-use jormungandr_integration_tests::common::load::ClientLoadError;
-use jormungandr_testing_utils::testing::node::{JormungandrRest, RestRequestGen};
-use jortestkit::{
-    load::Configuration,
-    load::Monitor,
-    prelude::{parse_progress_bar_mode_from_str, ProgressBarMode},
-};
+use crate::mjolnir_app::MjolnirError;
+use jormungandr_testing_utils::testing::node::explorer::load::ExplorerRequestGen;
+use jormungandr_testing_utils::testing::node::Explorer;
+use jortestkit::load::Configuration;
+use jortestkit::load::Monitor;
+use jortestkit::prelude::parse_progress_bar_mode_from_str;
+use jortestkit::prelude::ProgressBarMode;
 use structopt::StructOpt;
 use thiserror::Error;
 
-pub fn main() -> Result<(), RestLoadCommandError> {
-    RestLoadCommand::from_args().exec()
-}
-
 #[derive(Error, Debug)]
-pub enum RestLoadCommandError {
-    #[error("No scenario defined for run. Available: [duration,iteration]")]
-    NoScenarioDefined,
+pub enum ExplorerLoadCommandError {
     #[error("Client Error")]
-    ClientError(#[from] ClientLoadError),
+    ClientError(#[from] MjolnirError),
 }
 
 #[derive(StructOpt, Debug)]
-pub struct RestLoadCommand {
-    /// Prints nodes related data, like stats,fragments etc.
+pub struct ExplorerLoadCommand {
+    /// Number of threads
     #[structopt(short = "c", long = "count", default_value = "3")]
     pub count: usize,
     /// address in format:
-    /// /ip4/54.193.75.55/tcp/3000
+    /// 127.0.0.1:80
     #[structopt(short = "e", long = "endpoint")]
     pub endpoint: String,
 
@@ -34,7 +28,7 @@ pub struct RestLoadCommand {
     #[structopt(short = "p", long = "pace", default_value = "2")]
     pub pace: u64,
 
-    /// amount of delay [seconds] between sync attempts
+    /// load duration
     #[structopt(short = "d", long = "duration")]
     pub duration: u64,
 
@@ -51,11 +45,11 @@ pub struct RestLoadCommand {
     pub measure: bool,
 }
 
-impl RestLoadCommand {
-    pub fn exec(&self) -> Result<(), RestLoadCommandError> {
-        let mut rest_client = JormungandrRest::new(self.endpoint.clone());
-        rest_client.disable_logger();
-        let mut request_gen = RestRequestGen::new(rest_client);
+impl ExplorerLoadCommand {
+    pub fn exec(&self) -> Result<(), ExplorerLoadCommandError> {
+        let mut explorer = Explorer::new(self.endpoint.clone());
+        explorer.disable_logs();
+        let mut request_gen = ExplorerRequestGen::new(explorer);
         request_gen.do_setup(Vec::new()).unwrap();
 
         let config = Configuration::duration(
@@ -65,7 +59,7 @@ impl RestLoadCommand {
             self.build_monitor(),
             0,
         );
-        let stats = jortestkit::load::start_sync(request_gen, config, "rest load test");
+        let stats = jortestkit::load::start_sync(request_gen, config, "Explorer load test");
         if self.measure {
             assert!((stats.calculate_passrate() as u32) > 95);
         }
